+++ arch/sh/kernel/process_32.c
@@ -167,7 +167,7 @@ void prepare_to_copy(struct task_struct *tsk)
 
 asmlinkage void ret_from_fork(void);
 
-int copy_thread(unsigned long clone_flags, unsigned long usp,
+int copy_thread(unsigned long long clone_flags, unsigned long usp,
 		unsigned long unused,
 		struct task_struct *p, struct pt_regs *regs)
 {
